requirements.txt
# Основные пакеты для анализа данных (обновленные до последних стабильных версий)
pandas==2.2.1            # Последняя LTS-версия Pandas для Python 3.11
numpy==1.26.4            # Совместим с pandas 2.2.1
polars==0.20.7           # Высокопроизводительный DataFrame (актуальная версия)

# Логирование (современные альтернативы)
loguru==0.7.2            # Удобное логирование
structlog==24.1.0        # Структурированные логи (последняя версия)
python-json-logger==2.0.7  # JSON-форматирование

# Jupyter и инструменты разработки
notebook==7.1.2          # Последняя версия Jupyter Notebook
jupyterlab==4.1.6        # Jupyter Lab (опционально)
ipykernel==6.29.3        # Ядро для Python 3.11

# Работа с БД
psycopg[binary]==3.1.18  # Современный клиент PostgreSQL (для TimescaleDB)
sqlalchemy==2.0.29       # ORM (если потребуется)

# Дополнительные утилиты
python-dotenv==1.0.0     # Загрузка .env файлов
tqdm==4.66.2             # Индикаторы прогресса